[VB-67] Рефакторинг
